+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.3.0 - 2023-05-01
+
+### Added
+
+- `Innmind\Framework\Application::route()`
+- `Innmind\Framework\Http\To`
+
 ## 1.2.0 - 2023-04-29
 
 ### Added